pyproject.toml
[tool.poetry]
name = "storage-unit"
version = "0.1.0"
description = ""
authors = ["MuhammadHassan1998"]
readme = "README.md"
packages = [{include = "storage_unit"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
black = "23.3.0"
flake8 = "6.0.0"
python-dotenv = "1.0.0"
psycopg2 = "^2.9.5"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
six = "1.16.0"
django-cors-headers = "3.14.0"
xlrd = "2.0.1"
pandas = "2.0.1"
openpyxl = "3.1.2"
asgiref = "3.6.0"
click = "8.1.3"
django-leaflet = "0.29.0"
et-xmlfile = "1.1.0"
geographiclib = "2.0"
geopy = "2.3.0"
mccabe = "0.7.0"
mypy-extensions = "1.0.0"
numpy = "1.24.3"
packaging = "23.1"
pathspec = "0.11.1"
platformdirs = "3.5.0"
psycopg2-binary = "2.9.6"
pycodestyle = "2.10.0"
pyflakes = "3.0.1"
pyjwt = "2.6.0"
python-dateutil = "2.8.2"
pytz = "2023.3"
sqlparse = "0.4.4"
tomli = "2.0.1"
tzdata = "2023.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
